HandGunから遠距離武器の求通部分を切り出し
