+++ Assets/SSP/Scripts/Weapons/HandGunModel.cs
@@ -1,0 +1,8 @@
+﻿using UnityEngine;
+
+public class HandGunModel : LongRangeWeaponModel
+{
+    public GameObject bullet;
+    public float bulletVelocity = 10;
+    public float bulletDeathTime = 5;
+}
