+++ Assets/SSP/Scripts/Weapons/HandGunModel.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: f0b914f142c9ed5438ce56c2ebc9a5af
+timeCreated: 1510670214
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
