+++ Assets/SSP/Scripts/Weapons/LongRangeWeapon.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 39db80289456d85419d89fb0dfba907e
+timeCreated: 1510669095
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
