+++ Assets/SSP/Scripts/Weapons/LongRangeWeaponModel.cs
@@ -4,8 +4,6 @@ using UnityEngine;
 
 public class LongRangeWeaponModel : WeaponModel
 {
-    public GameObject bullet;
-    public float bulletVelocity = 10;
-    public float bulletDeathTime = 5;
     public float coolTime = 0.5f;
+    public float scopeMagnification = 1.0f;
 }
